2024.06.13 snr : 설정시간 대 혼잡도페이지와 연결
